# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"微服务","version":""},"host":"localhost:8000","basePath":"/api","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/user/get_captcha_image":{"post":{"tags":["User"],"summary":"获取图形验证码","description":"获取图形验证码","operationId":"User#GetCaptchaImage","responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserGetCaptchaImageResponseBody","required":["errcode","errmsg"]}},"400":{"description":"Bad Request response.","schema":{"$ref":"#/definitions/UserGetCaptchaImageBadRequestResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/UserGetCaptchaImageInternalServerErrorResponseBody"}}},"schemes":["http"]}},"/user/login_by_username":{"post":{"tags":["User"],"summary":"使用账号密码登录","description":"使用账号密码登录","operationId":"User#LoginByUsername","parameters":[{"name":"LoginByUsernameRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/UserLoginByUsernameRequestBody","required":["username","password"]}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserLoginByUsernameResponseBody","required":["errcode","errmsg"]}},"400":{"description":"Bad Request response.","schema":{"$ref":"#/definitions/UserLoginByUsernameBadRequestResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/UserLoginByUsernameInternalServerErrorResponseBody"}}},"schemes":["http"]}},"/user/update_password":{"post":{"tags":["User"],"summary":"修改登录密码","description":"修改登录密码\n\n**Required security scopes for jwt**:\n  * `role:user`","operationId":"User#UpdatePassword","parameters":[{"name":"Authorization","in":"header","description":"JWT used for authentication","required":true,"type":"string"},{"name":"UpdatePasswordRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/UserUpdatePasswordRequestBody","required":["old_password","new_password"]}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserUpdatePasswordResponseBody","required":["errcode","errmsg"]}},"400":{"description":"Bad Request response.","schema":{"$ref":"#/definitions/UserUpdatePasswordBadRequestResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/UserUpdatePasswordInternalServerErrorResponseBody"}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]}}},"definitions":{"CaptchaResponseBody":{"title":"CaptchaResponseBody","type":"object","properties":{"captchaId":{"type":"string","description":"验证码ID","example":"Vel at quas deserunt quia atque."},"image":{"type":"string","description":"图片base64","example":"Esse voluptates et rerum assumenda."}},"example":{"captchaId":"Soluta distinctio doloremque eaque.","image":"Delectus reiciendis eum qui quasi."},"required":["image","captchaId"]},"CredentialsResponseBody":{"title":"CredentialsResponseBody","type":"object","properties":{"expires_in":{"type":"integer","description":"有效时长（秒）：生成之后x秒内有效","example":25200,"format":"int64"},"token":{"type":"string","description":"JWT token","example":"eyJhbGciOiJIUz..."}},"example":{"expires_in":25200,"token":"eyJhbGciOiJIUz..."},"required":["token","expires_in"]},"SessionResponseBody":{"title":"Mediatype identifier: session; view=default","type":"object","properties":{"credentials":{"$ref":"#/definitions/CredentialsResponseBody"},"user":{"$ref":"#/definitions/UserResponseBody"}},"description":"会话 (default view)","example":{"credentials":{"expires_in":25200,"token":"eyJhbGciOiJIUz..."},"user":{"id":"Doloremque animi itaque aperiam vero voluptas ut.","mobile":"Vel explicabo est adipisci architecto.","username":"Est eaque."}},"required":["user","credentials"]},"SuccessResultResponseBody":{"title":"Mediatype identifier: successresult; view=default","type":"object","properties":{"ok":{"type":"boolean","description":"success","example":true}},"description":"成功信息 (default view)","example":{"ok":true},"required":["ok"]},"UserGetCaptchaImageBadRequestResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"GetCaptchaImage_bad_request_Response_Body result type (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"UserGetCaptchaImageInternalServerErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"GetCaptchaImage_internal_server_error_Response_Body result type (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"UserGetCaptchaImageResponseBody":{"title":"UserGetCaptchaImageResponseBody","type":"object","properties":{"data":{"$ref":"#/definitions/CaptchaResponseBody"},"errcode":{"type":"integer","description":"错误码","example":0,"minimum":0,"maximum":999999},"errmsg":{"type":"string","description":"错误消息","example":""}},"example":{"data":{"captchaId":"Sequi ad recusandae harum sunt culpa aliquid.","image":"Eos quae rerum provident."},"errcode":0,"errmsg":""},"required":["errcode","errmsg"]},"UserLoginByUsernameBadRequestResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"LoginByUsername_bad_request_Response_Body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"UserLoginByUsernameInternalServerErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"LoginByUsername_internal_server_error_Response_Body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"UserLoginByUsernameRequestBody":{"title":"UserLoginByUsernameRequestBody","type":"object","properties":{"captchaId":{"type":"string","description":"图形验证码ID","default":"","example":"5l","minLength":1,"maxLength":128},"humanCode":{"type":"string","description":"图形验证码","default":"","example":"bvx","minLength":4,"maxLength":8},"password":{"type":"string","description":"密码","example":"password","minLength":1,"maxLength":128},"username":{"type":"string","description":"用户名","example":"user","minLength":1,"maxLength":128}},"example":{"captchaId":"l3","humanCode":"9ub","password":"password","username":"user"},"required":["username","password"]},"UserLoginByUsernameResponseBody":{"title":"UserLoginByUsernameResponseBody","type":"object","properties":{"data":{"$ref":"#/definitions/SessionResponseBody"},"errcode":{"type":"integer","description":"错误码","example":0,"minimum":0,"maximum":999999},"errmsg":{"type":"string","description":"错误消息","example":""}},"example":{"data":{"credentials":{"expires_in":25200,"token":"eyJhbGciOiJIUz..."},"user":{"id":"Sit praesentium beatae dignissimos blanditiis voluptas ut.","mobile":"Repellat ea.","username":"Sed fugit sint ex nulla."}},"errcode":0,"errmsg":""},"required":["errcode","errmsg"]},"UserResponseBody":{"title":"Mediatype identifier: user; view=default","type":"object","properties":{"id":{"type":"string","description":"ID","example":"Quas deleniti expedita harum inventore."},"mobile":{"type":"string","description":"手机号","example":"Officiis similique qui."},"username":{"type":"string","description":"用户名","example":"Sint molestiae."}},"description":"用户","example":{"id":"Eaque voluptas.","mobile":"Et rem quo laboriosam.","username":"Modi odit."},"required":["id","username","mobile"]},"UserUpdatePasswordBadRequestResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"UpdatePassword_bad_request_Response_Body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"UserUpdatePasswordInternalServerErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"UpdatePassword_internal_server_error_Response_Body result type (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"UserUpdatePasswordRequestBody":{"title":"UserUpdatePasswordRequestBody","type":"object","properties":{"new_password":{"type":"string","example":"abc123","minLength":6,"maxLength":128},"old_password":{"type":"string","example":"123abc","minLength":1,"maxLength":128}},"example":{"new_password":"abc123","old_password":"123abc"},"required":["old_password","new_password"]},"UserUpdatePasswordResponseBody":{"title":"UserUpdatePasswordResponseBody","type":"object","properties":{"data":{"$ref":"#/definitions/SuccessResultResponseBody"},"errcode":{"type":"integer","description":"错误码","example":0,"minimum":0,"maximum":999999},"errmsg":{"type":"string","description":"错误消息","example":""}},"example":{"data":{"ok":true},"errcode":0,"errmsg":""},"required":["errcode","errmsg"]}},"securityDefinitions":{"jwt_header_Authorization":{"type":"apiKey","description":"JWT 认证\n\n**Security Scopes**:\n  * `role:user`: 用户\n  * `role:admin`: 管理员","name":"Authorization","in":"header"}}}